@@ -33,7 +33,7 @@
     "docs": "https://auth0.com/docs/customize/integrations/cms/wordpress-plugin"
   },
   "require": {
-    "php": "^7.4 || ^8.0",
+    "php": "^8.0",
     "ext-filter": "*",
     "ext-json": "*",
     "ext-mbstring": "*",
@@ -52,8 +52,9 @@
     "nyholm/psr7": "^1.5",
     "pestphp/pest": "^1.21",
     "php-http/mock-client": "^1.5",
+    "php-stubs/wordpress-stubs": "^6.0",
     "phpstan/phpstan": "^1.7",
-    "phpstan/phpstan-strict-rules": "^1.3",
+    "phpstan/phpstan-strict-rules": "^1.4.3",
     "phpunit/phpunit": "^9.5",
     "rector/rector": "^0.13.6",
     "squizlabs/php_codesniffer": "^3.7",
